moved methods to contextual controllers
